FORAI/requirements.txt
# FORAI Dependencies

# Core (required)
numpy>=1.20.0

# PDF reports (optional but recommended)
fpdf2>=2.7.0

# Progress bars
tqdm>=4.60.0

# Local LLM - llama-cpp-python (optional)
# llama-cpp-python>=0.2.0

# Local LLM - Ollama client (optional)
# requests>=2.28.0

# Neural networks - for CognitiveMesh (optional)
# torch>=2.0.0

# RL framework (optional)
# stable-baselines3>=2.0.0
# gymnasium>=0.29.0

# Graph database - Kuzu (optional, SQLite fallback available)
# kuzu>=0.1.0
